requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
google-auth==2.32.0
requests==2.32.3
